Onboarding takes /Welcome (nobodies-collective/Humans#1091)
WelcomeController + view + its nine Welcome_* copy keys (all six locales) move to
Humans.Onboarding; the active-member check goes through Base's RoleChecks instead of
Shell's claims transformation. Tests follow to Humans.Onboarding.Tests.
