+++ src/Humans.Base/Resources/SharedResource.de.resx
@@ -2378,33 +2378,6 @@
   <data name="AdminDetail_ManageEmails" xml:space="preserve">
     <value>E-Mails verwalten</value>
   </data>
-  <data name="Welcome_Title" xml:space="preserve">
-    <value>Willkommen</value>
-  </data>
-  <data name="Welcome_Heading" xml:space="preserve">
-    <value>Du bist dabei.</value>
-  </data>
-  <data name="Welcome_Subtitle" xml:space="preserve">
-    <value>Dein Ticket für Elsewhere 2026 ist bestätigt.</value>
-  </data>
-  <data name="Welcome_ShiftsTitle" xml:space="preserve">
-    <value>Jetzt machen wir das gemeinsam möglich.</value>
-  </data>
-  <data name="Welcome_ShiftsBody" xml:space="preserve">
-    <value>Elsewhere ist kein Festival, zu dem du einfach kommst — es ist eines, das du mitaufbaust. Jeder human wählt Schichten: Aufbau vorher, Betrieb währenddessen und Abbau danach.</value>
-  </data>
-  <data name="Welcome_ShiftsBody2" xml:space="preserve">
-    <value>Um Schichten zu sehen und dich anzumelden, brauchst du ein Humans-Konto. Das dauert nur ein paar Minuten.</value>
-  </data>
-  <data name="Welcome_CTA" xml:space="preserve">
-    <value>Anmelden oder Konto erstellen</value>
-  </data>
-  <data name="Welcome_Footer" xml:space="preserve">
-    <value>Schon ein Konto? Gleicher Button — melde dich an und du landest direkt bei den Schichten.</value>
-  </data>
-  <data name="Welcome_Contact" xml:space="preserve">
-    <value>Fragen? Schreib uns an</value>
-  </data>
   <data name="Profile_BuildStrip_Title" xml:space="preserve">
     <value>Build availability &amp; tracking</value>
     <comment>TODO: translate — copied from English</comment>
