+++ src/Humans.Base/Resources/SharedResource.fr.resx
@@ -2378,33 +2378,6 @@
   <data name="AdminDetail_ManageEmails" xml:space="preserve">
     <value>Gérer les e-mails</value>
   </data>
-  <data name="Welcome_Title" xml:space="preserve">
-    <value>Bienvenue</value>
-  </data>
-  <data name="Welcome_Heading" xml:space="preserve">
-    <value>C'est bon, t'es dedans.</value>
-  </data>
-  <data name="Welcome_Subtitle" xml:space="preserve">
-    <value>Ton billet pour Elsewhere 2026 est confirmé.</value>
-  </data>
-  <data name="Welcome_ShiftsTitle" xml:space="preserve">
-    <value>Maintenant, on fait ça ensemble.</value>
-  </data>
-  <data name="Welcome_ShiftsBody" xml:space="preserve">
-    <value>Elsewhere, c'est pas un festival où tu viens — c'est un que tu aides à construire. Chaque human choisit des créneaux : montage avant, opération pendant et démontage après.</value>
-  </data>
-  <data name="Welcome_ShiftsBody2" xml:space="preserve">
-    <value>Pour voir les créneaux et t'inscrire, il te faut un compte Humans. Ça prend juste quelques minutes.</value>
-  </data>
-  <data name="Welcome_CTA" xml:space="preserve">
-    <value>Se connecter ou créer un compte</value>
-  </data>
-  <data name="Welcome_Footer" xml:space="preserve">
-    <value>Déjà un compte ? Même bouton — connecte-toi et tu arrives directement sur les créneaux.</value>
-  </data>
-  <data name="Welcome_Contact" xml:space="preserve">
-    <value>Des questions ? Écris-nous à</value>
-  </data>
   <data name="Profile_BuildStrip_Title" xml:space="preserve">
     <value>Build availability &amp; tracking</value>
     <comment>TODO: translate — copied from English</comment>
