+++ src/Humans.Base/Resources/SharedResource.it.resx
@@ -2378,33 +2378,6 @@
   <data name="AdminDetail_ManageEmails" xml:space="preserve">
     <value>Gestisci email</value>
   </data>
-  <data name="Welcome_Title" xml:space="preserve">
-    <value>Benvenuto/a</value>
-  </data>
-  <data name="Welcome_Heading" xml:space="preserve">
-    <value>Ci sei.</value>
-  </data>
-  <data name="Welcome_Subtitle" xml:space="preserve">
-    <value>Il tuo biglietto per Elsewhere 2026 è confermato.</value>
-  </data>
-  <data name="Welcome_ShiftsTitle" xml:space="preserve">
-    <value>Ora facciamo succedere tutto insieme.</value>
-  </data>
-  <data name="Welcome_ShiftsBody" xml:space="preserve">
-    <value>Elsewhere non è un festival a cui partecipi — è uno che aiuti a costruire. Ogni human sceglie i turni: allestimento prima, operazione durante e smontaggio dopo.</value>
-  </data>
-  <data name="Welcome_ShiftsBody2" xml:space="preserve">
-    <value>Per vedere i turni e iscriverti, ti serve un account Humans. Ci vogliono solo un paio di minuti.</value>
-  </data>
-  <data name="Welcome_CTA" xml:space="preserve">
-    <value>Accedi o crea un account</value>
-  </data>
-  <data name="Welcome_Footer" xml:space="preserve">
-    <value>Hai già un account? Stesso pulsante — accedi e arrivi direttamente ai turni.</value>
-  </data>
-  <data name="Welcome_Contact" xml:space="preserve">
-    <value>Domande? Scrivici a</value>
-  </data>
   <data name="Profile_BuildStrip_Title" xml:space="preserve">
     <value>Build availability &amp; tracking</value>
     <comment>TODO: translate — copied from English</comment>
