+++ src/Humans.Base/Resources/SharedResource.resx
@@ -1058,15 +1058,6 @@
   <!-- Issues section -->
 
   <!-- Issues — controller status messages (TempData) -->
-  <data name="Welcome_Title" xml:space="preserve"><value>Welcome</value></data>
-  <data name="Welcome_Heading" xml:space="preserve"><value>You're in.</value></data>
-  <data name="Welcome_Subtitle" xml:space="preserve"><value>Your ticket for Elsewhere 2026 is confirmed.</value></data>
-  <data name="Welcome_ShiftsTitle" xml:space="preserve"><value>Now let's make this thing happen together.</value></data>
-  <data name="Welcome_ShiftsBody" xml:space="preserve"><value>Elsewhere isn't a festival you show up to — it's one you help build. Every human picks shifts: setting up before, keeping things running during, and taking it all down after.</value></data>
-  <data name="Welcome_ShiftsBody2" xml:space="preserve"><value>To browse and sign up for shifts, you need a Humans account. It takes a couple of minutes.</value></data>
-  <data name="Welcome_CTA" xml:space="preserve"><value>Sign in or create account</value></data>
-  <data name="Welcome_Footer" xml:space="preserve"><value>Already have an account? Same button — sign in and you'll land straight on the shifts page.</value></data>
-  <data name="Welcome_Contact" xml:space="preserve"><value>Questions? Reach us at</value></data>
 
   <!-- Google sync removal notifications (issue peterdrier/Humans#639) -->
