+++ src/Sections/Humans.Onboarding/OnboardingResource.de.resx
@@ -191,4 +191,31 @@
   <data name="Onboarding_ShiftsSignUpFailed" xml:space="preserve"><value>Anmeldung fehlgeschlagen.</value></data>
   <data name="Onboarding_ShiftsSignUpRangeFailed" xml:space="preserve"><value>Anmeldung für den Datumsbereich fehlgeschlagen.</value></data>
   <data name="Onboarding_ShiftsTitle" xml:space="preserve"><value>Wähle eine Schicht</value></data>
+  <data name="Welcome_Title" xml:space="preserve">
+    <value>Willkommen</value>
+  </data>
+  <data name="Welcome_Heading" xml:space="preserve">
+    <value>Du bist dabei.</value>
+  </data>
+  <data name="Welcome_Subtitle" xml:space="preserve">
+    <value>Dein Ticket für Elsewhere 2026 ist bestätigt.</value>
+  </data>
+  <data name="Welcome_ShiftsTitle" xml:space="preserve">
+    <value>Jetzt machen wir das gemeinsam möglich.</value>
+  </data>
+  <data name="Welcome_ShiftsBody" xml:space="preserve">
+    <value>Elsewhere ist kein Festival, zu dem du einfach kommst — es ist eines, das du mitaufbaust. Jeder human wählt Schichten: Aufbau vorher, Betrieb währenddessen und Abbau danach.</value>
+  </data>
+  <data name="Welcome_ShiftsBody2" xml:space="preserve">
+    <value>Um Schichten zu sehen und dich anzumelden, brauchst du ein Humans-Konto. Das dauert nur ein paar Minuten.</value>
+  </data>
+  <data name="Welcome_CTA" xml:space="preserve">
+    <value>Anmelden oder Konto erstellen</value>
+  </data>
+  <data name="Welcome_Footer" xml:space="preserve">
+    <value>Schon ein Konto? Gleicher Button — melde dich an und du landest direkt bei den Schichten.</value>
+  </data>
+  <data name="Welcome_Contact" xml:space="preserve">
+    <value>Fragen? Schreib uns an</value>
+  </data>
 </root>
